+++ src/main/java/ru/myrecord/front/data/dao/RoomDAO.java
@@ -8,7 +8,7 @@ import ru.myrecord.front.data.model.User;
 import java.util.List;
 
 @Repository("roomDAO")
-public interface RoomDAO extends JpaRepository<Room, Long> {
-	Room findById(Long id);
+public interface RoomDAO extends JpaRepository<Room, Integer> {
+	Room findById(Integer id);
 	List<Room> findByUserAndActiveTrueOrderByIdAsc(User user);
 }
