+++ src/main/java/ru/myrecord/front/service/iface/ServiceService.java
@@ -5,7 +5,7 @@ import ru.myrecord.front.data.model.User;
 import java.util.List;
 
 public interface ServiceService {
-    public Service findServiceById(Long id);
+    public Service findServiceById(Integer id);
     public List<Service> findByActive(User user);
     public void add(Service service);
     public void update(Service service);
